Fix Percona XtraDB Cluster guest to work with v5.6
The current implmentation of Percona clusters only supports
PXC 5.5. To make the implementation work with v5.6, some changes
need to be made to the base Mysql manager. The biggest change
is that we can no longer store the os_admin client credentials
in the server my.cnf file. This commit moves those credentials to
a .my.cnf file stored in the trove homedir. This will be read
by the guest when attempting sqlchemy connections to the local
mysql database and will also be picked up automatically by other
mysql clients (such as mysqladmin).

Also, since PXC 5.6 requires PXB 2.3, this commit fixes the issue
previously discovered with that combination. The previous fix was
to pin PXB to v2.2. The fix is to specify host=127.0.0.1 in the
client section of my.cnf.

Change-Id: Ib73b4f8ba40ddf211d0ec88069c4807186fce139
Parial-bug: 1532969