Fix issue of mismatched test-requirements.txt file
When you run devstack, the trove plugin installs python-troveclient by
forcing upper-constraints.txt in requirements to be changed as below.

-python-troveclient===2.0.0
+-e file:///opt/stack/python-troveclient#egg=python-troveclient

As a result, pinning test-requirements.txt to use a file for
python-troveclient to come from tarballs.openstack.org causes a
failure.

The error you will see is "Could not satisfy constraints for
'python-troveclient': installation from path or url cannot be
constrained to a version".

This change modifies test-requirements.txt and matches
global-requirements.txt, and devstack makes the change (above) to
upper-constraints.txt which accomplishes the same thing.

Change-Id: I3bb748a575d67f9f419f51837555d18ed5d7ca32
Closes-Bug: 1539818