Cleanup Mysql replication test on completion
The "Mysql" (e.g. mysql, percona, mariadb) scenario user tests
would fail if not run in isolation. It turns out that the
replication test which runs earlier wasn't cleaning itself up
properly and causing 2 failures in the user tests. This commit
adds code to cleanup the replication tests as they complete.
(Note: it is likely the replica slave users should be cleaned
up during detach_replica but this is another bug not addressed
by this commit [1])

[1] https://bugs.launchpad.net/trove/+bug/1549376

Change-Id: Ib512801a16999f4a5410a7caf45481dc8c30d010
Closes-bug: 1549387