Secure mongodb instances and clusters by default
MongoDB's default behaviour is to have the authorization rules disabled.
This means any client can access the database without credentials. This
is too large of a risk for default behaviour, so authorization needs to
be enabled by default.

This fix addresses the problem by setting the authorization parameter to
enabled for standalone instances, and specifying key files for cluster
instances to use.

Change-Id: I5895ce66ec450c3efce7a7e0f126ac13ac4c1b12
Closes-bug: #1507841