Revert "Time to get rid of most vestiges of slave_of"
This reverts commit 0c94a5bae0aacff7fde9a9742b7d526096e51767

Removing 'slave_of' from the instance create API was a
backward incompatible change.

The parameter should be removed if/when trove has a new
version in their API, or adopts microversions.

This change depends on the change which reverts the
removal of slave_of from python-troveclient since
the tests depend on that argument being available in the
client.

Note that the tests could later be updated to no longer
use 'slave_of' to avoid a deprecation warning from the
client.

Conflicts:
        trove/tests/api/replication.py

NOTE(mriedem): The conflict was due to
5515c5291cc0ebacad59973c87ba696b5ffc5efa.

Depends-On: I9b3ec6c101a89f30e61fc683808506c404e5474f

Change-Id: Ib6811293d99fbc20f6a52ee10de7c98bfb682b18
Partial-Bug: #1538506