Trove's tox tests should respect upper-constraints.txt
This will force pip install to use the upper-constraints.txt specified
version of pip modules. When you don't do this, you are out on the
bleeding edge and become unstable everytime some python library in the
wide world changes in a way that you don't expect.

The releasenotes job has no constraints file, so override the install
command there.

This commit is being proposed on stable/mitaka and therefore the
install command includes the clear designation of the branch. This is
because, on branches other than master, we need to be sure to get the
upper-constraints.txt file for that branch.

Change-Id: I8417e0fefae9711d0f71b8f97bbb0e4ea5f8ccf2
Closes-Bug: #1563038
(cherry picked from commit aca90168f52169afb36500d6af77723b1a707008)