Fixes migrations for MySQL 5.6.* and MariaDB 10.1.*
In MySQL 5.6/MariaDB 10.1 and later versions there was a change
on the restrictions over foreign keys (FK).

In 019-datastore-fix.py, we attempt to change a column that is part
of a FK from NULL to not NULL. This sort of modifications are not
longer allowed.

This situation caused trove-manage db_sync and trove-manage db_migrate
to fail.

To workaround this, the FK check before executing the ALTER query
is disabled and then reenabled.

Change-Id: I666d01235f2c3225aca3fe7520ebdf6d53831cab
Closes-Bug: #1473226
(cherry picked from commit 679e2283ecc888bd567955b4bf431c87043e8d12)