Mysql GTID replication fails when data inserted
If you have a master and a slave configured and you insert
new data into the master, it will cause subsequent replica
create to fail.

The problem is that we weren't setting the gtid_purged variable
using the metadata in the xtrabackup_binlog_info file. The
Mysql GTID replication strategy was adjusted to account for this.

A release note has been added.

The replication scenario tests were enhanced to validate this
issue.

Note: this issue doesn't occur with MariaDB GTID replication
because it mechanism is different.

Scenario tests were run succesfully on Mysql, Percona and
MariaDB with this change in place.

Change-Id: I66c8b6278afa50ba14e4bb7888e3a25dc657a9e4
Closes-bug: 1563574