encrypt_data(): don't encode IV to base64
Encoding the IV (Initialization vector) to base64 means that each
byte only contains 6 bits of entropy per byte, whereas an IV can
contain up to 8 bits of entropy of byte. The Crypto AES function
accepts any byte string for the IV.

Since the IV is included in the encrypted string (as a prefix of a
fixed size), this change is backward compatible. Data encrypted with
the old code is decodable with the old and with the new code.

Change-Id: I210a21b80599110309fb92b4670e7a5749e94756