Add New Relic License module driver
The recent addition of module support in Trove (see
https://blueprints.launchpad.net/trove/+spec/module-management ) does
not include the New Relic license module driver. This has been added.

A decorator to streamline writing drivers (by handling common errors)
was also added, and the ping driver modified to use it as well.

Since this code is dependent on having an image with New Relic
installed, no changes were made to the scenario tests with
respect to this new driver.

An addition flag was added to the 'apply' interface that passes in
whether a module was created with 'admin options.'  This allows
some rudimentary access control to be implemented.

Depends-On: I6fb23b3dbbec98de9ee1e2731bcfc56ab3c0ca42
Change-Id: I282cf533c99e351d23f3b86aae727ae4bf279b64
Closes-Bug: #1571711