Change recursion depth to '1' for mock detection
With the addition of more python3 tests, the py34 jobs is
failing intermittently due to timeouts.  It seems that python3
is significantly slower in running the unittests, and *way*
slower in doing the dangling-mock detection. Setting the
recursion value to '1' will make the tests run faster while
we figure out the best strategy going forward.

Change-Id: I922c9c465f0009be5ccd399982ff5f6c7fe02416