MySQL do not retrieve Password in get user
MySQL Manager also retrieves the hashed password when getting a user.
Aside from not being used or returned anywhere the value would be useless
anyways as it is only a hash of the real password.

Loading it from the user table only introduces compatibility issues with
more recent versions of MySQL that remove the Password 'column'.

This patch set removes the column from the retrieving query and leaves the
password on the returned user object unassigned.

Change-Id: I10852475f498698557074d432999720c21cb5799