Fix secure method to work with PXC
When PXC was implemented, the secure method was overridden as the
stock one was causing problems on prepare. This issue seems to
be caused by the order of work done in the call. If the admin
user is set up first and then used to delete the anon users,
everything works for all flavors of MySQL.

Tested using the scenario tests on MySQL, Percona, PXC and MariaDB.

Also cleaned up some of the common code and made the connection URL
more resilient to special password characters. Updated the 'supported'
tests and fixed the user actions one to work with PXC.

Change-Id: Idf9a096b8870e428090454528b79930f3d2fd760
Closes-Bug: #1585355