Fix MariaDB scenario replication test
When the binlog_format fix was put in for Percona and MySQL,
(see https://review.openstack.org/#/c/300253 )
MariaDB didn't need to be fixed. The scenario test was not set
up to handle the differences between the 3 flavors,
and has been fixed.

In a nutshell, MariaDB now expects to have 'STATEMENT' binlog
format, not 'MIXED' like the other two.

Change-Id: I1ced565ce18d5492126875c66fb74e85c091e15c
Closes-Bug: #1585384