Run Python 3 unit tests in parallel
Replace testtools.run test runner with os-testr for Python 3 to run
tests in parallel.

Replace also the whitelist with a blacklist to quickly identify tests
which still must be ported to Python 3.

Blacklist backup.test_backup_models: even if it worked when run with
testtools.run, it fails when run using os-testr. We can investigate
this issue later, it's more important to reduce the duration of the
py34 job to reduce the risk of random timeout.

Add a dependency to os-testr.

Partially implements: blueprint trove-python3
Change-Id: I5ea7ebf7eec15b75a5fed07aab7bd15e5e6cd6d6