Move the rabbit/rpc options to its own section
Using the DEFAULT section for rabbit configuration is deprecated, so
all rabbit related option should be moved to the oslo_messaging_rabbit
section. The example agent config still respects the
TROVE_HOST_GATEWAY option.

This change has to depend on a change in redstack that otherwise
overwrites the hard work we do here :)

Change-Id: Ic169efdb2eb453abb6c4401a082586715a84a038
Depends-On: Iec072877e9174e136b4f0c1c806d7f9b377a7f2e