Add Install Guide
This adds trove-specific contents of the OpenStack Installation Guide
in the trove repo per [1]. A separate change will remove the heat
contents from the OpenStack Installation Guide for Newton per [2].

This is a verbatim import from the content in openstack-manuals with
minimal changes to get content building.

Also adds tox.ini environment for install-guide and adds
openstackdocs-theme to test-requirements.txt.

From lmichaels -- added a statement to the next-steps doc.

[1]
http://specs.openstack.org/openstack/docs-specs/specs/newton/project-specific-installguides.html
[2]
http://specs.openstack.org/openstack/docs-specs/specs/newton/installguide.html
[3] https://review.openstack.org/#/c/314229/

Partially-Implements: blueprint projectspecificinstallguides
Change-Id: Idf8cb90536fd4b4d9a2d0695ef30f498b1283410