Use correct message for missing default datastore
The error message when trying to create an instance
with a non-existent default datastore was misleading:

ERROR: Datastore 'mysql' cannot be found. (HTTP 400)

This is now more descriptive:

ERROR: Please specify datastore. Default datastore 'mysql' cannot be
found. (HTTP 400)

If no default datastore is defined, the message is:

ERROR: Please specify datastore. No default datastore is defined. (HTTP
400)

If the user specified a datastore, then the regular error
message is returned.

A new exception DatastoreDefaultDatastoreNotDefined was created
and DatastoreDefaultDatastoreNotFound modified to take 'datastore'
as an argument.

Change-Id: I0b7856864bfecfe3173e42cf098ad759ec8334a2
Closes-Bug: #1331408