Have Trove plugin install dashboard
At present if the Trove plugin is enabled and Horizon is enabled,
the Trove dashboard still needs to be enabled (which is currently
done by redstack). This should be handled by the Trove plugin
itself.

With this fix, if Trove and Horizon are installed the dashboard
will be automatically set up.

Some variable names were modified slightly as well to bring
them in line with the rest.

The cleanup code was also moved to the 'stop' section, which is
where it probably belonged right from the beginning.

Depends-On: I71cf27a57e8d2cd141928f26ac01bd5ca14c723f
Change-Id: I6586801bff4282e3bbf11e45e4f171fffcab285e
Partial-Bug: #1593806