Fix config registration for Trove tempest plugin
The Trove tempest plugin was created in this commit:
Ibad5502cdca0e34cca4b5bc8a7106611f7393cf2. The tests however was
not working because the plugin specific configuration variables
was not being set correctly. This particular issue has been
fixed in this changeset. Using the following commands:
    tox -e all-plugin -- trove.tests.tempest.tests.api
4/5 tests pass. Will look into the failing test in a different
changeset.

Change-Id: Iad34209adb120e63cebbf12939baf30c5f017844
Partial-Bug: #1597857