Remove times.dbm file for each tox run
If you run the py27 tox tests then attempt to run the py34 tests,
the following error will occur:

    py34 runtests: commands[0] | ostestr
    --blacklist_file=blacklist-py3.txt
    db type could not be determined
    running=${PYTHON:-python} -m subunit.run discover
    ./trove/tests/unittests --list
    The test run didn't actually run any tests

This happens because the apparently the times.dbm file generated
by py27 is not compatible with py34. The converse is not true though -
if you run py34 first, py27 will work.

The work-around is to delete the file, which is what is now done
before running every tox test. Once the bug has been fixed, the
work-around can be removed.

This was originally noticed in python-troveclient, however it
also manifests in the Trove project.

coverage.xml was also added to the .gitignore file as it
was appearing in the git list of modified files.

Change-Id: I312b3879c764b7b803d35e0bb62f0f47291f6693
Closes-Bug: #1565928