Insulate TroveContext from oslo.context changes
Most of the recent issues with compatibility between running guests
and the control plane have been related to changes in oslo.context.
By putting some compatibility code into TroveContext, we should be
able to insulate the environment during upgrades.

Goal here is to reduce the chance of issues when the controller is
upgraded before the guests (which is what every operator wants
to do).

The basic idea here is to introspect the current TroveContext
including the RequestContext base class and remove any unknown
arguments coming off the wire before the constructor is called.

Change-Id: Ia4752c7dfe7bcfedebfaca9f13fb8bc6323d6001
Closes-bug: 1551468