Module re-apply does not reflect updated values
If you apply a module, remove it, update the module with a new name,
and apply it again, module-query will still show the old name.

This also applies to datastore, datastore_version and type.

All four attributes are now updated correctly and scenario
tests added.

When writing the scenario tests it became evident that resetting
a module from a datastore to all_datastores was broken too,
so this was fixed (along with all_datastore_versions).  This
fix required a change to the Trove client as well.

Change-Id: Ic2e9226bcd31c4a79d84ecb7941a47875eabd73d
Depends-On: I0f226f09db08f635c542b27d8d2f898050d327fa
Closes-Bug: #1611525
Closes-Bug: #1612430