Cleanup trove's tox.ini and .gitignore
This commit cleans up a couple of issues with Trove's tox.ini.

1. something that was reported to me by Matthew Treinish who
observed that the tox jobs are oddly bundled together.

This is important because infra's reporting is unable to generate some
useful metrics and tracking of trove's jobs because of the way they
are tied together.

2. with the cleanup of docs, apidocs is no longer a directory and can
be cleaned up, similarly api-ref is a thing and should be handled in
.gitignore.

3. that py27 used to use testr, switched it to using ostestr. I've
retained it for coverage.

4. the sqlite db needs to be deleted between runs because we otherwise
fail (on a developer configuration) because some tests intentionally
appear to leave stuff around.

5. I've set all ostestr jobs to run with --serial because there are
clearly cases where multiple tests can attempt to setup the DB at the
same time and get into trouble. Till we can fix those (if we feel it
is worthwhile) it is more reliable to run ostestr in --serial mode. It
only takes 2.5 minutes. This includes the cover jobs which runs with
--no-parallel.

An infra change will follow that will enable testing jobs.

Change-Id: I0a971c6ef6c1a4669490f2aa0e67c795fe86e190
Closes-Bug: 1618098
Closes-Bug: 1618149