Skip 'eject valid master' replication test
There is a test to eject a valid master during replication that is
failing more and more often in the scenario tests. Basically the
eject will only work if the heartbeat from the current master is
more than a minute old. This shouldn't happen in the scenario
test run, but it does - quite often.

Since this is consuming a large amount of gate resources, and the
bug isn't that onerous (but is probably hard to fix), the current
'eject valid master' test will be turned off until the fix lands.

The scenario test will print out the bug number during each run
as a reminder (using the new SkipKnownBug method).

Co-Authored-By: Peter Stachowski <peter@tesora.com>
Co-Authored-By: Amrith Kumar <amrith@tesora.com>
Author:     Peter Stachowski <peter@tesora.com>
Change-Id: Ia543da551ad4394d4964541f9db474e0792b9337
Related-Bug: #1622014
(cherry picked from commit 7d8d743d8e168f64fa893c975d7301507d4985d8)
Conflicts:
	trove/tests/scenario/runners/replication_runners.py
	trove/tests/scenario/runners/test_runners.py