Remove Trove's support for Heat
This is now master ... ocata has been branched.

It is important to clarify what this is because it is a confusing
statement. If you are using heat and want to use heat to provision a
trove instance, that will work. Heat can call the Trove API and things
will work.

On the other hand, if Trove wants to provision an instance, it will
not use Heat. This change eliminates all support for heat within
Trove; you can't use a heat template as the mechanism that Trove will
use to provision resources.

Depends-On: I2f2a12207581a94fb8561a6d65a3a79b4a29b063

Change-Id: I964cdd3ae67261b6fd7f666b7795b35e12788c9c