Fix Redis cluster grow scenario tests
Cluster tests were enabled for Redis as they
should be working now.  The tests were disabled for
MariaDB (and PXC) until the galera issues are sorted out.

Support for neutron was added in cluster tests by
making sure that the proper nic is passed in on
cluster cre
ate/grow tests.

The changes for client non-caching was also implemented.

The scenario tests were also broken apart into
sections to facilitate easier manual testing.

Support for cluster upgrade (currently in review)
was also added so that when datastore implementations
are pushed up they can be tested immediately.

Change-Id: I637cbe30fa42528e8abd2a301baa2ec5048b61fc
Closes-Bug: 1645775