PostgreSQL guest_log test fails intermittently
When testing the user log for PostgreSQL, it occasionally fails due
to the fact that its heartbeat actually performs a select on the
underlying database. This can sometimes cause new lines to be written
to the log file while the file is being published. This will return a
state of 'Partial' instead of 'Published.'

Both these states are now valid for this test.

See:
http://logs.openstack.org/98/356698/3/check/
  gate-trove-scenario-functional-dsvm-postgresql-nv/
  d45bf5a/console.html#_2016-10-12_00_01_21_419989

This fix should make the gate more stable as it is proposed that
PostgreSQL be made voting.

Change-Id: I32b182148073d8a0bcd2792cdc730ab684c3e760
Closes-Bug: #1633603