Fail on deleting non-existing database
Check that the database exists before deleting it and
fail if it does not.

Trove database API implementation closely followed MySQL
behavior which is to silently ignore deletes on non-existing
databases.
This strategy however no longer works that well with
other datastores which generally fail.

Trove users should be given a consistent experience.
Failing with an explanatory message is safer with respect to other
datastores and it is also consistent with how other Trove API
(user, backup, cluster) work.

Note: guestagent implementation was fix to handle
      limit=None in database (and user) queries.
      N >= None always returns True

Change-Id: I1dc26c7b40d6e4867b145b8470e6e5277e5dcdb4
Closes-Bug: 1293954