allow PATH_TROVE to be specified by gate or defaulted by trovestack
When run as redstack, which was in a different repository, PATH_TROVE
could be set as $DEST/trove unambiguously. However, now that
trovestack is in the trove repository, this is no longer the case. The
gate will have trove in a different location than the rest of the
repositories that get cloned.

Therefore we have to handle this situation by defaulting PATH_TROVE to
be based on the location of trovestack with the option to override.

Also use a better way to find TROVESTACK_SCRIPTS (than pwd).

Change-Id: Ic788c772017fbb410ce65939c90986e2d400e36d