Add test for flavor
We increased in three patch for flavor-list disk, vcpu, ephemeral,
but did not have a corresponding test to test whether can obtain.
we once have a test by trove/api-ref/source/samples/db-flavors-response.json,
but is not have now, even if I had deleted the file, the test still
can get through it.

Change-Id: Ibe93c975d93158407fefaeeb2e7a48325d03d5b7