Improve list-of-ports validation
List options tcp_ports and udp_ports are lists of strings and some
with a '-' in the middle to indicate a range. To help validate the
options better a new type was introduced to oslo.config called Range.

oslo.config version 3.18.0 merged this Range type which will no
longer require the following in our project:
    * utility function gen_ports because a Range of ints are returned
    * test to check for proper from-to format, the type is smart so
    it flips the numbers around for us. So 63000-300 returns
    Range(300, 63001) 63001 because inclusion=True is set by default.

Change-Id: I63b6a865a3f3c79202dd299f6cd25dd59e182252
Closes-Bug: #1500141