clean up references to trove-integration and redstack
Our readmes and rst files had references to trove-integration and
redstack. This commit cleans up those references.

Jian Song had also been trying to make some of these kinds of changes
to the documentation and to reduce conflicts and merges, I'd requested
him to hold off. As a result, I think it is only fair that I list him
as a co-author.

Change-Id: I685659391fd04eefd250e72d789da5a0d7b7aace
Co-Authored-By: jiansong <jian.song@easystack.cn>