Allow splitting test groups into sub groups
When registering a new test group we
give it a group name and a list of tests to run.

    register(['foo'], [t1, t2, t3])

creates group 'foo' running tests t1, t2 and t3.

This method has been extended to allow passing a dict
of tests with keys used as group name suffixes.

    register(['foo'], sub1=[t1, t2], sub2=[t3])

creates group 'foo' running all tests t1, t2 and t3
and groups 'foo_sub1' running t1, t2
and group 'foo_sub2' running t3

The plan is to split long-running datastore tests
into multiple sub-groups that can run as separate jobs.

Generally, we will want to run replication and
clustering in one job separate from everything else.
I have registered a 'single' (tests on a single instance)
and 'multi' (tests woring on a set of instances) sub-groups
for all datastores (for consistency).

This does not change how the tests run.
It just creates '*_single' and '*_multi' sub-groups in addition
to the existing '<datastore>_supported' groups.
We will need to update the project config to run sub-jobs
for datastores that require it.

Change-Id: Iea2a996043b4c2d1889995e160891af33ffab5b6