Fix slow nodes causing guest_log to fail
The gate is more often than not using very slow nodes.
This is causing failures in the guest_log in that more data
can be available to publish 'immediately' after a publish.
Some of these were fixed but others are now causing problems.

All checks for this are now set to include 'Partial' as a
valid state to avoid these remaining race conditions.

Change-Id: I736072d769b7b97082d1dc34d4b0cc567454deb7
Closes-Bug: 1646246