Tweak trove devstack plugin
Tweak the trove devstack plugin to force KVM if possible. This change
is semi-experimental. I would like to see how it performs in the gate,
it seems to be fine on my local machine. It takes some precautions to
only force kvm if it is possible.

In speaking with infra, there are also some known stability issues
with kvm and part of the experiment here is to see what those are. The
other part is to quantify the improvement that it brings to the trove
gate.

Change-Id: I209d3202b4f6414cd3430c11f2ccdb8d5b541d94