formalizing configure_nova_kvm
The earlier hack has proved to be quite good, and has not produced
many problems. Messages echoed to the summary log are not stashed in
logstash, moving to a single message in the log instead. Changed the
function name to a more formal name (configure_nova_kvm) instead of
tweak_nova.

Change-Id: I85364c6530058e964a8eba7fb515d7deadfd5d72