Replace pexpect with processutils in volume.py
In investigating the issues with resize, it was noticed that pexpect
is used throughout guestagent/volume.py. These were replaced with
oslo_concurrency.processutils.execute to allow more robust error
handling/reporting.

Certain cases were made more tolerant to unknown states (for
example resize_fs now makes sure the volume is unmounted).
This (hopefully) will fix the intermittent gate issues
plaguing volume-resize lately.

While examining the errors returned from the guest, it was
observed that in one case a fault message from the guest was
almost immediately overwritten by a less-detailed one
from the taskmanager.  Handling was put in place for this
case to only write the fault if the previous fault was 'old'
enough.

The unit tests for volume were cleaned up as well with
regards to proper mocking and checking that the calls
were actually made properly.

Depends-On: Ia30cf5826415e47d59e6d407f95e4d7ff4e8ff6c
Change-Id: I844dd26a968a50532a6220a0762a302c2cadc2d0
Closes-Bug: 1651589