Remove negative backup tests from api run
These tests have been failing more regularly as we sometimes have very
fast machines. Basically the issue is the the backup completes before
the negative tests finish, resulting in an error that looks like:

FAIL: proboscis.case.MethodTest
(test_restore_instance_from_not_completed_backup)

These tests are handled sufficiently in the scenario tests that they
don't need to run here anymore and as such were removed. The
scenario tests are more stable in that they pump data in the database
that causes the backup to take longer and increasing the window of
success for these tests.

Change-Id: I8bae94f8794ba259a76c2cba7598f697aae7e1cb
Closes-Bug: #1651794