Trove user-delete API can use periods in user names in fact
I can create and delete a user named "mytest.abc" correctly both
with mysql-5.6 and mongodb-3.0, not as the details describe that
"Do not use periods in user names".

In addition, in order to create a mongodb database user in trove,
its format must be as "DBName.UserName". And, I can delete the users correctly!

Closes-Bug: #1652889
Change-Id: I2316943475b0bcfdb498733e221d2021b8a827a0