Install Redis 3.2.6 by compilation
The version of Redis installed by the current elements
uses a ppa that is stalled at 3.0.7.  This seems to
have some issues that are fixed in later versions.

The recommended method of installing Redis is to compile
the code yourself, so the elements have been changed
to do this.

The version was bumped to 3.2.6 (latest stable) however
the Redis cluster tests were not reenabled as there still
seems to be an issue with them.

New config options were added to the template; these
may need to be updated to make sure all the new ones
are there.

Change-Id: I1a604b15ade815cdb51c5484d4285e504508a6c4
Partial-Bug: 1652964