Skip redis clustering tests
Redis clustering support is currently unstable because of a
known transient bug in the redis server (CLUSTERDOWN)
fixed in version 3.2

The PPA we have been using to install Redis in our images
does not include this version.

We deactivate Redis clustering test in this patch set to
stabilize the gate.

Change-Id: I0987b0c2a92e796a511832d420ed5b7d3c7779b8