Disable wait for force-delete
In the scenario tests, sometimes the force-delete command doesn't end up
deleting the instance. This seems to be a timing issue in that it
doesn't occur during every run (and therefore might be hard to track
down). It's not terribly critical as you can run the command again and
it will probably work the second time, but until it's fixed the test
has been disabled.

Change-Id: Id5b041cb4d5276c5015921784526423e766d032c
Related-Bug: 1656422