fix requirements file branching for trove guests
When trove-integration was a thing, and was branchless, and not part
of the trove repository, we needed this kludge to get the guest
requirements. Now that trove-integration is no longer a thing, and the
image build is branched and part of the trove repository, we don't
need to keep requirements files which are version stamped.

The elements now just pick up either fedora-requirements.txt or
ubuntu-requirements.txt

This change will merge in master and get backported into stable/ocata.

Change-Id: I87b887befb2d12f607ffec919c79d462fbde42ca
(cherry picked from commit 5529c368e4788ab9a224970bae9076c015a3da12)