we now need to have dib installed explicitly
dib is now going to have to be installed explicitly for us to actually
run the gate. the old method of just cloning the repo won't work with
dib v2.

There are other ways to accomplish this but I'm choosing this approach
knowing that it isn't ideal. First, what's the downside? This means
that dib gets installed wherever trove control plane is installed. Is
that necessary, NO. Is it the end of the world, NO.

What are the options?

- do what Octavia did and have a second requirements.txt in a folder
  called diskimage-create. Would this work, yes but we don't have such
  a mechanism right now and nothing would keep this requirements.txt
  file in sync with g-r that I know of.

- hardcode the pip install command somewhere in the devstack
  plugin. yes, this would work but this would also have no ability to
  sync with g-r.

So, while the solution proposed here isn't ideal, it strikes me as not
bad, and much less risk than the other two options.

We also need to figure out where the elements are and fix the path for
disk-image-create.

Some pep8 failures appears to have crept in as well, those are
addressed.

Change-Id: Ic64d91e082102057417995a0f6851d03b9e0ca74