Fixing PROXY_AUTH_URL not being populated properly
In a recent patch some imports were added to the top of
cmd files, before the call to load the configuration
was issued. This caused anything initilized before the
configuration to fallback on default values.

Because of this, for example PROXY_AUTH_URL found in
trove/common/remote.py would always be set to the
default value: "http://0.0.0.0:5000/v2.0".
> PROXY_AUTH_URL = CONF.trove_auth_url

This patch changes the config values in the
trove/common/*remote.py files to be lazy loaded.

Co-Authored-By: Riccardo Pittau <rpittau@blizzard.com>
Change-Id: Ie3893680c406d92fe2b20e892c29fce675d7f224
Closes-Bug: #1671655