Fix the gate; py27 job failure with failure to import designate tests
The issue here appears to be that while unittests2 iterates the
directories and fines the directory dns, and then goes to try and
import the test called test_designate_driver, something else in the
path is showing up ahead of the tests and causing the search for
dns.test_designate_driver to fail. Renaming the directory is a kludge,
but it works.

And in the grand scheme of things, I'm good with this for now.

Change-Id: I30db9d15002ef86ef3b257c95298d54e766696c1