Improve Gate: address multiple issues relative to timing
This commit merges two commits, 447750[1] and 448630[2] in the fond
hope that this will improve the odds of its being able to merge.

The log operation test assumes that the log will be Ready but in some
cases, Partial is also a valid result. Add that into the list of
allowed results.

In addition, this change aims to support cases where test suites last
much more than 1 hour and where a token can expire unexpectedly mid
operation. This is by no means a 'fix' for the bug here, just a way to
get the gate to be more stable.

[1] https://review.openstack.org/447750
[2] https://review.openstack.org/448630

Change-Id: Ib888dd55f04952139508a7cb0fee60e192e1b77c
Closes-Bug: 1674520
Partial-Fix: 1674510