fix issue with Nova passthrough deprecation
Nova Networking was deprecated in Newton and some of the API calls
that were part of NN were still in use by Trove. With the arrival of
python-novaclient 8.0.0, this caused a failure.

Nova also deprecated the image pass through capability. The management
API uses some of that to test whether an image id is valid or
not. This change fixes the management API, and the attendant scenario
tests.

Depends-On: I142f97d691fa55e9824714c9c224f998ad72337e

Change-Id: I2f2a12207581a94fb8561a6d65a3a79b4a29b063
Closes-Bug: #1690936