How to create a trove instance
For the longest time I've had a hard time understanding why my instances
were in ERROR or stuck in BUILD state. This guide fills in the
missing details that are necessary for setting up a trove instance.

I hope this guide will expand into other debugging steps/tips but
this is a good start.

I often found myself reserving too many resources that weren't
available. I didn't realize the virtual machine hogs so much of
the resources, or at least on my machine that is.

This guide will help those who are new to the world of cloud
computing and give them an understanding of how a trove instance
relates to your system's hardware.

Change-Id: I2dcfadb77b2a66f7e4ba3363aa54c8ad89605a88