Log the right attached configuration id
Method 'rolling_configuration_update' in trove/cluster/models.py.
It allow re-applying the same configuration to an instance.In if code
block, if the instance does not attach any configuration or it has
attached the configuration that the parameter passed, then do attach
thing. "else" block means the instance has attached another config,
should use "instance.configuration.id" instead of "configuration_id",
one is a attached configuration, one is a wanted configuration, and
they are not the same.
Besides, fix a spelling mistake in trove/instance/models.py

Change-Id: I4b6fb338de611ce39a8173dcf4b6e17714c380b9
Closes-Bug: 1693466