Update ubuntu cassandra to supported repo and version
The datastax cassandra repository is publicly flagged as abandoned,
and java 7 is no longer available in modern releases of Ubuntu.

This moves the cassandra repo from the unsupported datastax repo
to the official Apache repo. This also updates to use Java 8 since
7 is no longer supported.

Both the cassandra release to be used and the Java used by this
element can be changed by setting the environment variables
CASSANDRA_RELEASE and CASSANDRA_JAVA accordingly. Notes in the
elements install script.

Change-Id: I3afc56c6c1a5443c6b71bab8f47222100584833a