Use DocumentedRuleDefault instead of RuleDefault
The DocumentedRuleDefault class has advantages over the RuleDefault
class because it allows project developers the opportunity to
describe how policies map to actual APIs. This makes things a lot
easier for operators trying to understand policies for a given
project, and doesn't require them to dig in the code to figure out
what they protect.

Polices are also splitted into small files for easier maintainance.

This is part of an effort to move policy and documentation of policy
into code, treating it like configuration.

bp policy-and-docs-in-code

Co-Authored-By: Zhao Chao <zhaochao1984@gmail.com>
Change-Id: Ifcdbccae7350ec69ad7d92197b9bfca8aeb8352b