Enable integration tests
Re-enable the integration tests.

Integration tests were temporarily disabled under
9ec134996a1162de0d74ef915dc138bbacbcfdf1 to avoid the
circular dependency with python-troveclient and allow
the merging of the python-troveclient changes necessary
to complete the fixing of the gates.

This also increases the TROVE_AGENT_CALL_HIGH_TIMEOUT to
help the upgrade test pass. The upgrade test has an agent call
that takes longer than the current timeout when running in
the gate on Nova compute with virt_type set as qemu.

Change-Id: I8459cd64cdf6107e537f4e7c793c3626fa4bf9f9