Fix gate issues
NOTE(mriedem): This is a squash of four commits.

The first commit message is:

Jobs run as jenkins on Zuul v2 but run as user zuul on Zuul v3.

NOTE(mriedem): This is only taking one specific fix from the large
change on master (queens) since we don't need the other changes
for stable/pike to fix the gate.

(cherry picked from commit 9ec134996a1162de0d74ef915dc138bbacbcfdf1)

This is the 2nd commit message:

Update .gitreview for stable/pike

This is the 3rd commit message:

Update UPPER_CONSTRAINTS_FILE for stable/pike

This is the 4th commit message:

Increase the TROVE_AGENT_CALL_HIGH_TIMEOUT to
help the upgrade test pass. The upgrade test has an agent call
that takes longer than the current timeout when running in
the gate on Nova compute with virt_type set as qemu.

NOTE(smatzek): This is only taking one fix from this change as
the other is N/A in this case as it was enabling the disabled
integration tests.

(cherry picked from commit d7b773d2b07f0c95909ef73c4644a63f09fd82c3)

Change-Id: Ic55187b0d73d4c572d7f8332882b4f455a6177c8