Fix integration test exception handling
Under 9ec134996a1162de0d74ef915dc138bbacbcfdf1 the exception
handling of the test_runners was changed to do a bare
raise of the original test failing exception. This was
done to preserve the original traceback. This had the
unintended side effect of allowing exceptions in the error
handling and log reporting code to be raised as the test
case failing exception.

This commit fixes the exception handling to save off the
exc_info and re-raise the original exception with original
traceback.

Change-Id: I13bcf194bb289749ec289216f2541d4923f5f2bf