Enable longer Keystone token life
The devstack plugin sets a 3 hour Keystone token life
but Keystone is already running when this configuration change
is called so Keystone still runs with the default for the tests.

Add a restart of Keystone to enable the longer life.

Change-Id: Ie340b1c0b56652fc9ca4c0981468363c1be9fb25
(cherry picked from commit a3514d84208d0244bb3a8aa3fc76092349c40115)