Enhance test case fail to build  message
When instances fail to build successfully during the gate runs
it is common to see a KeyError on 'ip'.  This is coming from the
test_runner code which is expecting the instance to have an IP
address.

This puts a check on 'ip' before its usage and fails the tests
with an more explicit error message.

Change-Id: Ibd926e90190024a5b1979b3a09fa1eda12aa7106