Fix requirepass problem with redis
When we use configuration to configure the request for the redis
requirepass, Since we did not rebuild the adminclient with
password, the previous client did not have permission to ping.
Will cause the status to enter shutdown. It is now modified to
rebuild the client as long as the configuration changes are made

co-author-by:Fan Zhang <zh.f@outlook.com>
Change-Id: I9a17eaf2b69da0571295b38c5b2ec87dbc30e1d9
Closes-Bug: #1708376