Fix qemu image compatibility mode
The qemu compatibility mode preserves compatibility with the
previous generation of the qemu-img tools. In 2014, when Ubuntu
Trusty was introduced, it was set to compat=0.10 to provide
compatibility with servers installed with Ubuntu Precise as
both Trusty and Precise were supported at that time.

Precise is now End of Life since April 2017, so the compatibility
mode is being updated to the default mode that exists for Trusty
which is compat=1.1.  The mode has not been updated since then.

Updating the mode has the advantage of picking up support for
newer qemu features like more efficient copy-on-read behavior.
Setting it to an explicit value ensures that the mode is not
silently changed.

Change-Id: Ifdfccf252585e7e2805b6dfc0940b630ad746cfb