Enable other Ubuntu architectures to utilize diskimage-builder
This is accomplished by parameterizing the -a arch argument
which is currently hardcoded to amd64.

For Ubuntu distributions, the package architecture as reported by
Ubuntu will be passed into disk-image-create via the -a argument.
This field controls which packages are installed by DIB.

Note diskimage-builder invokes commands in a chroot'd
environment, so cross architecture builds are not supported
by definition. The platform architecture of the target
image is always the same as the caller's.

Change-Id: Ie91209681b2291cc7a21895ad21dd1af5e82f32a