Allow the user to disable tmpfs when building guest images
This is accomplished through the use of a new environment
variable USE_TMPFS.  The default action remains to use tmpfs.

If the user wishes to disable the use of tmpfs, the
environment variable should be set as shown below before
invoking Trove to build an image.

export USE_TMPFS=false

Note tmpfs uses half of the available free memory for a
RAM based file system. This can lead to errors in the image
building process, for example when large compilations are
performed. In these cases, the user can try again with tmpfs
disabled simply by setting an environment variable. The
alternative is to provision another server or VM that has
more memory which is not always possible or desired.

Change-Id: I0f990d7866755ccf0537354752cd602fe0582ea0