Utilize Ubuntu's hardware enablement stack for trovestack
This brings in the latest fixes which is often important for
nested virtualization.  It is required for ppc64el.  This stack
is installed via a new post-install.d script in the trove
integration element ubuntu-xenial-guest.

A user that is manually creating datastore images may install
this stack by setting the following environment variable prior
to running the disk-image-create command.

export DIB_USE_HWE_KERNEL=true

Change-Id: Iaf01c543ea210bd4725ab909aff6884df451b198