Don't refresh trove code in guestagent once installed
Presently, the code is reloaded from the controller each time that
the guestagent is restarted.  The test in the trove-guest.service
file is wrong as it is looking for a directory as opposed to a file.

This improves the debug ability of the guestagent by allowing
LOG.debug statements to be manually inserted in the code.

Change-Id: I9703420d5ffd2b4e2d45dcd02435627e731250da