Add support for MySQL 5.7 on Ubuntu Xenial
Ubuntu Trusty continues to get MySQL 5.6.

The selection of the MySQL version is based on the version
of Ubuntu that is installed.  If the script trovestack is
invoked on Xenial, the version that is chosen is 5.7.  If
it is invoked on Trusty, then version 5.6 is chosen.

The only thing that is eliminated is the dubious combination
of MySQL 5.6 on Ubuntu Xenial.  This combination is probably
not supported by Ubuntu as it is down level from their
standard offering is probably not tested anywhere.

Note Xenial does not provide the 5.6 version of mysql.  This
was accomplished by initializing the Apt repo with the
Trusty software repository.

The size of the root file system is increased from 3 to 4 GBs
as the Mysql Xenial image does not fit in 3 GBs. This has an
impact on the flavors that are used by Trove for testing as the
name of the flavor includes the size of the root file system.

This is turn caused a change to each of the db specific test
config files as the trove falvors are referenced in them.

Change-Id: I4f4e497208b8f4728580e48239a8ae208e0a96dd