Improve Vertica 9.x support
During the Vertica installation, the install script try to get AWS
metadata even if --ignore-aws-instance-type is specified.

The only way to avoid the _get_macs error is to define
--ignore-install-config option.

During the cluster creation, if the process is too long, Vertica
ask a question:
  Do you want to continue waiting? (yes/no) [yes]

From the adminTools documentation there is no non-interactive or
timeout options. The only way to have this question answered is
to run "echo yes | " before the "adminTools -t create_db" command.

The only available version from Vertica website is 9.0.1.

Closes-Bug: #1745493
Change-Id: I94472082fac606f4ff14f26c2ca2b620ef0a3bbb