Unable to grow/shrink Vertica 9.x cluster
During the Vertica cluster growing/shrinking,
the "update_vertica" command tries to get AWS
metadata even if --ignore-aws-instance-type is
specified.

The only way to avoid the "_get_macs error" is to
define --ignore-install-config option.

This is very similar to this bug[1] but this time
this for the "trove cluster-grow" and "trove
cluster-shrink" commands.

[1] https://bugs.launchpad.net/trove/+bug/1745493

Change-Id: I35f2035039f72b3aa2d9beb665096d4fb57e1c4c
Closes-Bug: #1747050