Fix client recreation in Redis root-disable test
In the Redis root-disable scenario test, when the root user is disabled,
a check for ensuring the old password is not valid for connecting to
Redis sever fails because connected Redis clients are not affected by
the change of the 'requirepass' parameter, they can still issue commands
like 'PING'. So when a password is provided, a new client should always
be created.

Change-Id: I4e948a20f2b67e49c36bac5c20f411d0b84969ef