Skip root state inherting scenario tests for Redis
Redis instances does not inherit root state from backups, and this
feature is also not suitable for it, because Redis actually does not
have a user account management system. The "root password" for Redis
instance is just a configuration parameter for the underlying server.
So all scenario tests about checking root state inheritance should be
skipped.

Further investigations should also be done to the root API for Redis
datastore: we may have to delete the RootHistory database record when
disabling root, in this way the user could tell the underlying Redis
server is not enabling AUTH currently, this could make less confusion
for the end users. However if we're going this direction, we should
also make it clear that the RootHistory records mean different things
for different datastore types.

Change-Id: I988fcd906e074e591ccc1c60f09139df94814b5b